+++ src/main/kotlin/blackjack/ui/output/PlayingView.kt
@@ -23,7 +23,6 @@ object PlayingView {
     }
 
     fun showDealerHit() {
-        println()
         println("딜러는 16이하라 한장의 카드를 더 받았습니다.")
         println()
     }
