+++ src/test/kotlin/blackjack/domain/participant/DealerSpecs.kt
@@ -1,5 +1,6 @@
-package blackjack.domain
+package blackjack.domain.participant
 
+import blackjack.domain.CustomDeck
 import blackjack.domain.card.Card
 import blackjack.domain.card.Denomination.FIVE
 import blackjack.domain.card.Denomination.KING
@@ -9,11 +10,9 @@ import blackjack.domain.card.Denomination.TWO
 import blackjack.domain.card.Suit.DIAMOND
 import blackjack.domain.card.Suit.HEART
 import blackjack.domain.card.Suit.SPADE
-import blackjack.domain.participant.BlackJack
-import blackjack.domain.participant.Bust
-import blackjack.domain.participant.Dealer
-import blackjack.domain.participant.Deck
-import blackjack.domain.participant.Stay
+import blackjack.domain.cards
+import blackjack.domain.hand
+import blackjack.domain.to
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.equality.shouldBeEqualToComparingFields
