+++ src/test/kotlin/blackjack/domain/participant/StateSpecs.kt
@@ -1,13 +1,9 @@
-package blackjack.domain
+package blackjack.domain.participant
 
 import blackjack.domain.card.Point
-import blackjack.domain.participant.BlackJack
-import blackjack.domain.participant.Bust
-import blackjack.domain.participant.Hittable
 import blackjack.domain.participant.Match.DRAW
 import blackjack.domain.participant.Match.LOSE
 import blackjack.domain.participant.Match.WIN
-import blackjack.domain.participant.Stay
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.inspectors.forAll
